+++ packages/client-ledger/src/sdk/ledgers.ts
@@ -132,9 +132,8 @@ export class Ledgers extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }
 
@@ -581,9 +580,8 @@ export class Ledgers extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }