+++ packages/client-ledger/src/sdk/transactions.ts
@@ -235,9 +235,8 @@ export class Transactions extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }