+++ packages/client-ledger/RELEASES.md
@@ -77,3 +77,13 @@ Based on:
 - [typescript v1.0.0-alpha.7] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.7] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.7 - packages/client-ledger
+
+## 2024-07-18 00:15:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.340.0 (2.375.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.8] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.8] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.8 - packages/client-ledger